pyproject.toml
[project]
name = "kgpaper"
version = "0.1.0"
description = "Paper Knowledge Graph Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "rdflib>=7.0.0",
    "streamlit>=1.30.0",
    "streamlit-option-menu>=0.3.0",
    "st-cytoscape>=0.0.5",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
